fix: auth middlewares applied on all categories routes

+++ server/database/models/CategoryRepository.js
@@ -14,11 +14,11 @@ class CategoryRepository extends AbstractRepository {
     return rows;
   }
 
-  async readByCategoryId(categoryId) {
+  async readByCategoryId(category) {
     const [rows] = await this.database.query(
       `
-      SELECT * FROM ${this.table} WHERE id = ?`,
-      [categoryId]
+      SELECT * FROM ${this.table} WHERE id = ? AND user_id = ?`,
+      [category.id, category.user_id]
     );
     return rows;
   }
@@ -34,16 +34,16 @@ class CategoryRepository extends AbstractRepository {
 
   async update(category) {
     const [result] = await this.database.query(
-      `UPDATE ${this.table} SET name = ?, icon = ? WHERE id = ?`,
-      [category.name, category.icon, category.id]
+      `UPDATE ${this.table} SET name = ?, icon = ? WHERE id = ? AND user_id = ?`,
+      [category.name, category.icon, category.id, category.user_id]
     );
     return result.affectedRows;
   }
 
-  async delete(id) {
+  async delete(category) {
     const [result] = await this.database.query(
-      `DELETE FROM ${this.table} WHERE id = ?`,
-      [id]
+      `DELETE FROM ${this.table} WHERE id = ? AND user_id = ?`,
+      [category.id, category.user_id]
     );
     return result.affectedRows;
   }
